+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Rename `Error::InvalidAlpnRepsonse` to correct a typo ([#612](https://github.com/fastly/Viceroy/pull/612))
 - Upgrade to Rust 1.95 ([#604](https://github.com/fastly/Viceroy/pull/604))
 - Add options for experimenting with wasm gc and exceptions. ([#601](https://github.com/fastly/Viceroy/pull/601))
+- Improve TLS certificate loading, handling and validation ([#478](https://github.com/fastly/Viceroy/pull/478))
 
 ## 0.16.5 (2026-03-23)